Multiple plugin templates (#6) merge in main
* Extended the original to include the option of creating diagnostics plugins

* renamed diagnostics to postprocessor

* Update cookiecutter.json

* Fixed a minor bug and improved the displayed defaults

* updated Python versions to match Pysteps

* Created guide for adding new plugin types/subtypes

* Update Extend_Template.ipynb

* Update test_bake_project.py

* Update macosx test to use the latest gcc versions like pysteps does.

* Reinstate macosx test, update action versions and direct to correct branch in pysteps main repo to test postprocessor_plugin.

* Update checks and rm commands in hooks: valid plugins are importer and diagnostic (not postprocessor)

This is more consistent with the structure of pysteps which has
io/importers
postprocessing/diagnostics

* Adapt json file to new plugin types.

* Change references to postprocessor-diagnostic to diagnostic only

* Update docs

* Update references to diagnostic[s]

* Update docs in Extend_template.ipynb.

* Fix syntax in json file.

* remove ensemblestat from VALID_PLUGIN_TYPES in hooks/post_gen_project.py
- currently we do not have ensemblestats plugins
- importer and diagnostic remain as plugin types

* Fix diagnostic plugin __init__.py

* Match expected naming as expected in pysteps repo
- added plugin_function as parameter (needed for importer plugins)
- updated setup script for entry_points
- updated function names

* Remove cookiecutter import from __init__

* Fix module imports and __init__ files - cookiecutter finally working :)

* plugin README template installation instruction corrected

* Minor correction to run tests

* Update test scripts with correct defaults and avoid obsolete functions in setuptools

* Update test script inside cookiecutter template

correct importer prefix and remove ensemblestats for now.

* Make sure that the tests not only bake the plugin but also install it before testing.

* remove ensemblestat from json file plugin types

* Add tests for diagnostics plugins, too.

---------

+++ tests/test_bake_project.py
@@ -1,13 +1,13 @@
 """
 Tests for the Pysteps Plugin template.
 """
+
 import shlex
 from contextlib import contextmanager
 
 import datetime
 import os
 import subprocess
-import sys
 from cookiecutter.utils import rmtree
 
 
@@ -26,17 +26,21 @@ def inside_dir(dirpath):
 
 
 @contextmanager
-def bake_in_temp_dir(cookies, *args, **kwargs):
+def bake_in_temp_dir_and_install(cookies, *args, **kwargs):
     """
     Delete the temporal directory that is created when executing the tests
     :param cookies: pytest_cookies.Cookies,
         cookie to be baked and its temporal files will be removed
     """
     result = cookies.bake(*args, **kwargs)
+
     try:
+        run_inside_dir("pip install ./", str(result.project))
         yield result
     finally:
-        rmtree(str(result.project))
+        if result.project is not None and os.path.exists(result.project):
+            os.system("pip uninstall -y " + result.project.basename)
+            rmtree(str(result.project))
 
 
 def run_inside_dir(command, dirpath):
@@ -56,8 +60,8 @@ def check_output_inside_dir(command, dirpath):
 
 
 def test_year_compute_in_license_file(cookies):
-    with bake_in_temp_dir(cookies) as result:
-        license_file_path = result.project.join('LICENSE')
+    with bake_in_temp_dir_and_install(cookies) as result:
+        license_file_path = result.project.join("LICENSE")
         now = datetime.datetime.now()
         assert str(now.year) in license_file_path.read()
 
@@ -71,85 +75,93 @@ def project_info(result):
 
 
 def test_bake_with_defaults(cookies):
-    with bake_in_temp_dir(cookies) as result:
+    with bake_in_temp_dir_and_install(cookies) as result:
         assert result.project.isdir()
         assert result.exit_code == 0
         assert result.exception is None
 
         found_toplevel_files = [f.basename for f in result.project.listdir()]
-        assert 'setup.py' in found_toplevel_files
-        assert 'pysteps_importer_abc' in found_toplevel_files
-        assert 'tox.ini' in found_toplevel_files
-        assert 'tests' in found_toplevel_files
+        assert "setup.py" in found_toplevel_files
+        assert "pysteps_importer_institution_name" in found_toplevel_files
+        assert "tox.ini" in found_toplevel_files
+        assert "tests" in found_toplevel_files
 
 
 def test_bake_and_run_tests(cookies):
-    with bake_in_temp_dir(cookies) as result:
+    with bake_in_temp_dir_and_install(cookies) as result:
         assert result.project.isdir()
-        assert run_inside_dir('python setup.py test', str(result.project)) == 0
+        assert run_inside_dir("pytest", str(result.project)) == 0
+
+def test_bake_diagnostic_and_run_tests(cookies):
+    with bake_in_temp_dir_and_install(
+        cookies, extra_context={"plugin_type": "diagnostic"}
+    ) as result:
+        assert result.project.isdir()
+        assert run_inside_dir("pytest", str(result.project)) == 0
 
 
 def test_bake_withspecialchars_and_run_tests(cookies):
     """Ensure that a `full_name` with double quotes does not break setup.py"""
-    with bake_in_temp_dir(
-            cookies,
-            extra_context={'full_name': 'name "quote" name'}
+    with bake_in_temp_dir_and_install(
+        cookies, extra_context={"full_name": 'name "quote" name'}
     ) as result:
         assert result.project.isdir()
-        assert run_inside_dir('python setup.py test', str(result.project)) == 0
+        assert run_inside_dir("pytest", str(result.project)) == 0
 
 
 def test_bake_with_apostrophe_and_run_tests(cookies):
     """Ensure that a `full_name` with apostrophes does not break setup.py"""
-    with bake_in_temp_dir(
-            cookies,
-            extra_context={'full_name': "O'connor"}
+    with bake_in_temp_dir_and_install(cookies, extra_context={"full_name": "O'connor"}) as result:
+        assert result.project.isdir()
+        assert run_inside_dir("pytest", str(result.project)) == 0
+
+def test_bake_diagnostic_plugin(cookies):
+    with bake_in_temp_dir_and_install(
+        cookies, extra_context={"plugin_type": "diagnostic"}
     ) as result:
         assert result.project.isdir()
-        assert run_inside_dir('python setup.py test', str(result.project)) == 0
+        assert result.exit_code == 0
+        assert result.exception is None
 
+        found_toplevel_files = [f.basename for f in result.project.listdir()]
+        assert "setup.py" in found_toplevel_files
+        assert "pysteps_diagnostic__name" in found_toplevel_files
+        assert "tox.ini" in found_toplevel_files
+        assert "tests" in found_toplevel_files
 
 def test_bake_selecting_license(cookies):
     license_strings = {
-        'MIT license': 'MIT ',
-        'BSD license': 'Redistributions of source code must retain the ' +
-                       'above copyright notice, this',
-        'ISC license': 'ISC License',
-        'Apache Software License 2.0':
-            'Licensed under the Apache License, Version 2.0',
-        'GNU General Public License v3': 'GNU GENERAL PUBLIC LICENSE',
+        "MIT license": "MIT ",
+        "BSD license": "Redistributions of source code must retain the "
+        + "above copyright notice, this",
+        "ISC license": "ISC License",
+        "Apache Software License 2.0": "Licensed under the Apache License, Version 2.0",
+        "GNU General Public License v3": "GNU GENERAL PUBLIC LICENSE",
     }
     for license, target_string in license_strings.items():
-        with bake_in_temp_dir(
-                cookies,
-                extra_context={'open_source_license': license}
+        with bake_in_temp_dir_and_install(
+            cookies, extra_context={"open_source_license": license}
         ) as result:
-            assert target_string in result.project.join('LICENSE').read()
-            assert license in result.project.join('setup.py').read()
+            assert target_string in result.project.join("LICENSE").read()
+            assert license in result.project.join("setup.py").read()
 
 
 def test_bake_not_open_source(cookies):
-    with bake_in_temp_dir(
-            cookies,
-            extra_context={'open_source_license': 'Not open source'}
+    with bake_in_temp_dir_and_install(
+        cookies, extra_context={"open_source_license": "Not open source"}
     ) as result:
         found_toplevel_files = [f.basename for f in result.project.listdir()]
-        assert 'setup.py' in found_toplevel_files
-        assert 'LICENSE' not in found_toplevel_files
-        assert 'License' not in result.project.join('README.rst').read()
+        assert "setup.py" in found_toplevel_files
+        assert "LICENSE" not in found_toplevel_files
+        assert "License" not in result.project.join("README.rst").read()
 
 
 def test_using_pytest(cookies):
-    with bake_in_temp_dir(
-            cookies,
-            extra_context={'use_pytest': 'y'}
-    ) as result:
+    with bake_in_temp_dir_and_install(cookies, extra_context={"use_pytest": "y"}) as result:
         assert result.project.isdir()
-        test_file_path = result.project.join(
-            'tests/test_pysteps_importer_abc.py'
-        )
+        test_file_path = result.project.join("tests/test_pysteps_importer_institution_name.py")
         lines = test_file_path.readlines()
         # Test the new pytest target
-        assert run_inside_dir('python setup.py pytest', str(result.project)) == 0
+        assert run_inside_dir("pytest", str(result.project)) == 0
         # Test the test alias (which invokes pytest)
-        assert run_inside_dir('python setup.py test', str(result.project)) == 0
+        assert run_inside_dir("pytest", str(result.project)) == 0
